chore: LinkedList 타입을 Model로 이동 #2
